feat($breadcrumb): allow chain of states customization
The property 'ncyBreadcrumbParent' in states configuration override the parent state (only for the breadcrumb).
It is useful when 2 states are not parent in states configuration but must be parent in the breadcrumb.
For example, in ui-router sample : contact.list and contact.detail are configured as siblings but must be parent/child in breadcrumb :
contact.list > contact.detail

Closes #7