fix: organize state-level options in `ncyBreadcrumb` key instead of `data`
Originally, state-level options was defined in the `data` key which is the
key reserved for custom usages (according to UI-router documentation). It
causes problems because of data inheritance (#30). The ugly
`$$isStateDataProperty` function was checking if a property was in state's
data is its own.

Now, these options are defined in the `ncyBreadcrumb` key. They are no
longer inherited between states and `$$isStateDataProperty` is no longer
needed.

Closes #30
BREAKING CHANGE: state-level options has been moved under the custom key
`ncyBreadcrumb` in state's configuration.

To migrate the code follow the example below:
```
// Before
$stateProvider.state('A', {
  url: '/a',
  data: {
    ncyBreadcrumbLabel: 'State A'
  }
});
```

```
// After
$stateProvider.state('A', {
  url: '/a',
  ncyBreadcrumb: {
    label: 'State A'
  }
});
```
See API reference on wiki for more informations.