feat(ncyBreadcrumbLast): Add a new directive rendering the last step
This directive renders the last step of the breadcrumb (generally
the current state).

The directive is an attribute with an optional value. For example:
`<title ncy-breadcrumb-last></title>`

The value can be a short template (as ngBindTemplate) with the same
variables in scope as the ncyBreadcrumb directive:
- ncyBreadcrumbLabel
- ncyBreadcrumbLink
For example:
`<span ncy-breadcrumb-last="<a href='{{ncyBreadcrumbLink}}'>{{ncyBreadcrumbLabel}}</a>"></span>`

If no value is specified, the directive uses the default template `{{ncyBreadcrumbLabel}}`.

Closes #37