feat(ncyBreadcrumbLast): add custom templating
Add ability to set a custom template or templateUrl for
ncyBreadCrumbLast. Before, if a custom template that allowed HTML via
ngBindHtml, breadcrumb last would not behave properly since it escaped
HTML by default. A value could be assigned to the ncyBreadcrumbLast
directive to customize the template, but it would have to be done on
each instance of the directive used. Now, a template can be set once via
the templateLast and templateLastUrl properties of the configuration
object passed to $breadcrumbProvider.setOptions.